+++ src/sp_resources/src/control_box.rs
@@ -2,7 +2,7 @@ use sp_domain::*;
 use sp_runner::*;
 use std::collections::HashMap; // todo: macro depends on this...
 
-pub fn make_control_box(name: &str) -> Resource {
+pub fn create_instance(name: &str) -> Resource {
     resource! {
         name: name,
         command!{
@@ -12,8 +12,8 @@ pub fn make_control_box(name: &str) -> Resource {
             blue_light : bool,
         },
         measured!{
-            topic: "state",
-            msg_type: "control_box_msgs/msg/State",
+            topic: "measured",
+            msg_type: "control_box_msgs/msg/Measured",
 
             blue_light_on : bool,
         },
@@ -50,7 +50,7 @@ mod test {
     #[test]
     #[serial]
     fn test_control_box() {
-        let control_box = make_control_box("cb");
+        let control_box = create_instance("cb");
         println!("{:#?}", control_box);
 
         let bl = control_box
