+++ src/sp_resources/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod camera;
+pub mod control_box;
+pub mod dorna;
+pub mod gripper;
